refactor(share): remove barrel re-export, update import sites directly
