--- src/components/ShareButton.tsx
@@ -1,1 +1,0 @@
-export { ShareButton } from "./share/ShareButton";
